refactor: factorize pydantic term instantiation and resolve circular dependency
