--- src/cmipld/__init__.py
@@ -1,10 +1,0 @@
-from pydantic import BaseModel
-
-from cmipld.api.data_descriptors import DATA_DESCRIPTOR_CLASS_MAPPING
-
-
-def get_pydantic_class(data_descriptor_id_or_term_type: str) -> type[BaseModel]:
-    if data_descriptor_id_or_term_type in DATA_DESCRIPTOR_CLASS_MAPPING:
-        return DATA_DESCRIPTOR_CLASS_MAPPING[data_descriptor_id_or_term_type]
-    else:
-        raise ValueError(f"{data_descriptor_id_or_term_type} pydantic class not found")
