+++ src/cmipld/api/data_descriptors/__init__.py
@@ -27,6 +27,13 @@ from cmipld.api.data_descriptors.variable import Variable
 from cmipld.api.data_descriptors.variant_label import VariantLabel
 
 
+def get_pydantic_class(data_descriptor_id_or_term_type: str) -> type[BaseModel]:
+    if data_descriptor_id_or_term_type in DATA_DESCRIPTOR_CLASS_MAPPING:
+        return DATA_DESCRIPTOR_CLASS_MAPPING[data_descriptor_id_or_term_type]
+    else:
+        raise ValueError(f"{data_descriptor_id_or_term_type} pydantic class not found")
+
+
 DATA_DESCRIPTOR_CLASS_MAPPING: dict[str, type[BaseModel]] = {
     "activity": Activity,
     "consortium": Consortium,
