+++ src/cmipld/db/project_ingestion.py
@@ -8,7 +8,6 @@ from pydantic import BaseModel
 
 import cmipld.db as db
 import cmipld.settings as settings
-from cmipld import get_pydantic_class
 from cmipld.db import DBConnection, items_of_interest, read_json_file
 from cmipld.db.models.project import Collection, Project, PTerm
 import cmipld.db.universe_ingestion as universe_ingestion
